% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recuento.R
\name{recuento}
\alias{recuento}
\title{Calcula el valor Recuento de los últimos 30 minutos de macs}
\usage{
recuento()
}
\value{
json
}
\description{
Calcula el valor Recuento de los últimos 30 minutos de macs
}
\examples{
 recuento()

}
