% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempoPromedio.R
\name{tiempoPromedio}
\alias{tiempoPromedio}
\title{Calcula el valor promedio del último día}
\usage{
tiempoPromedio()
}
\value{
json
}
\description{
Calcula el valor proemdio del último día
}
\examples{
 tiempoPromedio()

}
